requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.5
Faker==2.0.1
idna==2.8
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
